adaptive dtype

+++ python/llm/src/bigdl/llm/transformers/models/bloom.py
@@ -75,7 +75,7 @@ def bloom_layer_norm_forward(self, hidden_states):
         if result.nelement != 0:
             return result
     input_dtype = hidden_states.dtype
-    result = F.layer_norm(hidden_states.to(torch.bfloat16),
+    result = F.layer_norm(hidden_states.to(self.weight.dtype),
                           self.normalized_shape, self.weight, self.bias, self.eps)
     return result.to(input_dtype)
